Adds support for multiple databases on one host
Each database has to be separated by a space in the config file and has to have
a corresponding username and password. Values are positional so that 3
databases need 3 usernames and 3 passwords (even if they are the same). First
database will use first username and first password etc.
This will be improved later.